NiceCropImageFunctionApp/requirements.txt
requests
azure-functions
numpy
opencv-python
azure-storage-blob